Cast ROI from config file to tuple
JSON convert python tuples to lists. This caused a bug where a tuple was compared to a list which resulted in "False" when the result should have been "True".

Specifically, this bug caused the `utils.bg_read` function to average over an ROI area when it shouldn't, which in turn made recOrder's BG correction incorrect (or at least different from expected).