Accomodate `waveorder`'s `Polscope_bg_correction` requirements
`waveorder`'s `Polscope_bg_correction` function requires a background to be supplied, and this background will be subtracted for all `bg_option` choices ('global', 'local_fit', and 'none').

Before this commit, `local_fit` background correction was applying a background correction from file (which might be empty if the GUI wasn't present, or it might be a real file if read from a config file) before proceeding to the local_fit correction.

After this commit, the 'local_fit' option is applying an "identity correction" (i.e. no correction) then proceeding to apply the local_fit correction.